requirements.txt
aiohttp==3.9.2
#BeautifulSoup4
fake-useragent==1.5.1
fastapi==0.110.1
uvicorn==0.29.0
SQLAlchemy==2.0.31    #agent code
#pydantic==2.7.4
pydantic==2.9.2     #agent code
python-jose==3.3.0
python-multipart==0.0.9
cryptography==36.0.1
#azure-core==1.30.1
azure-core==1.30.2     #agent code
databricks-sql-connector==3.4.0
#azure-cosmos==4.5.1
#azure-functions
azure-search==1.0.0b2     #agent code
#azure-search-documents==11.6.0b1
azure-search-documents==11.6.0b3     #agent code
#azure-storage-blob==12.19.1
#azure-storage-queue==12.9.0
#openai==1.16.2
openai==1.55.3     #agent code
python-dotenv==1.0.1
bcrypt==4.1.3
pyodbc==5.1.0
aiofiles==24.1.0
Pillow==10.3.0
validators==0.31.0

jinja2==3.1.4

#-------------agent code
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.10
langchain-openai==0.1.8
langchain-text-splitters==0.2.2
tiktoken==0.7.0
#streamlit==1.35.0   # Why this is needed
pyautogen==0.2.28
#pandas==2.2.2
#urllib3==2.2.2
#websockets==12.0
#openpyxl==3.1.5
sqlparse==0.5.0
azure-common==1.1.28
plotly==5.14.1
#tenacity==8.4.2
#python-dateutil==2.9.0.post0
matplotlib==3.9.2
seaborn==0.13.2
pydantic_core==2.23.4